requirements.txt
aiohttp==3.8.1
pytz==2022.1
selenium==4.3.0
unicode_slugify==0.1.5
webdriver_manager==3.5.4
